Домашнее задание по теме «Компоненты Spring Boot приложений»

+++ src/main/java/homeworks/homework21/src/main/java/tvapp/entity/Television.java
@@ -4,6 +4,8 @@ package tvapp.entity;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Getter
 @Setter
@@ -14,13 +16,13 @@ import lombok.*;
 public class Television {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     @Column(name = "brand")
     private String brand;
     @Column(name = "model")
     private String model;
     @Column(name = "price")
-    private Double price;
+    private BigDecimal price;
     @Column(name = "screen_size")
     private Integer screenSize;
